+++ twitter4j-core/src/main/java/twitter4j/StatusStreamImpl.java
@@ -95,7 +95,7 @@ class StatusStreamImpl extends AbstractStreamImplementation implements StatusStr
                 ((StatusListener)listener).onDeletionNotice(new StatusDeletionNoticeImpl(deletionNotice.getJSONObject("status")));
             }else{
                 JSONObject directMessage = deletionNotice.getJSONObject("direct_message");
-                ((UserStreamListener)listener).onDeletionNotice(ParseUtil.getInt("id", directMessage)
+                ((UserStreamListener)listener).onDeletionNotice(ParseUtil.getLong("id", directMessage)
                         , ParseUtil.getInt("user_id", directMessage));
             }
         }
