+++ twitter4j-core/src/main/java/twitter4j/UserStreamListener.java
@@ -31,7 +31,7 @@ package twitter4j;
  * @since Twitter4J 2.1.3
  */
 public interface UserStreamListener extends StatusListener {
-    void onDeletionNotice(int directMessageId, int userId);
+    void onDeletionNotice(long directMessageId, int userId);
 
     /**
      * @param friendIds
